Move default implementation of configuration to test only
