+++ src/lib.rs
@@ -2,12 +2,17 @@ pub mod packs;
 
 #[cfg(test)]
 pub(crate) mod test_util {
+    use configuration::Configuration;
     use packs::parsing::ruby::zeitwerk::get_zeitwerk_constant_resolver;
+    use std::collections::{HashMap, HashSet};
     use std::path::PathBuf;
 
     use packs::configuration;
 
+    use crate::packs::configuration::from_raw;
     use crate::packs::pack::Pack;
+    use crate::packs::raw_configuration::RawConfiguration;
+    use crate::packs::walk_directory::WalkDirectoryResult;
     use crate::packs::{self, constant_resolver::ConstantResolver};
 
     pub const SIMPLE_APP: &str = "tests/fixtures/simple_app";
@@ -50,6 +55,22 @@ pub(crate) mod test_util {
             enforce_privacy: Default::default(),
             enforce_visibility: Default::default(),
             enforce_architecture: Default::default(),
+        }
+    }
+
+    impl Default for Configuration {
+        fn default() -> Self {
+            let default_absolute_root = std::env::current_dir().unwrap();
+            let walk_directory_result = WalkDirectoryResult {
+                included_files: HashSet::new(),
+                included_packs: HashSet::new(),
+                owning_package_yml_for_file: HashMap::new(),
+            };
+            from_raw(
+                &default_absolute_root,
+                RawConfiguration::default(),
+                walk_directory_result,
+            )
         }
     }
 }
